# ServerSide/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Laxmi Enterprise API",
    "description": "Centralized Attendance and Vehicle Management Service",
    "version": "2.0"
  },
  "paths": {
...
    "/api/v1/employees/": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "List Employees",
        "operationId": "list_employees",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeResponse"
                  },
                  "title": "Response List Employees"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Create Employee",
        "operationId": "create_employee",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vehicles/": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "List Vehicles",
        "operationId": "list_vehicles",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Vehicle Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleResponse"
                  },
                  "title": "Response List Vehicles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Create Vehicle",
        "operationId": "create_vehicle",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{vehicle_number}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get Vehicle",
        "operationId": "get_vehicle",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Update Vehicle",
        "operationId": "update_vehicle",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Create Session",
        "operationId": "create_session",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceSessionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Date"
            }
          },
          {
            "name": "shift",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shift"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceSessionResponse"
                  },
                  "title": "Response List Sessions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assignments/sessions/{session_id}/vehicles/{vehicle_id}": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "summary": "Get Vehicle Assignments",
        "operationId": "get_vehicle_assignments",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleAssignmentResponse"
                  },
                  "title": "Response Get Vehicle Assignments"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttendanceSessionCreate": {
        "properties": {
          "sessionDate": {
            "type": "string",
            "title": "Sessiondate"
          },
          "shift": {
            "type": "string",
            "title": "Shift"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "in_progress"
          },
          "supervisorId": {
            "type": "string",
            "title": "Supervisorid"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "sessionDate",
          "shift",
          "supervisorId"
        ],
        "title": "AttendanceSessionCreate"
      },
      "AttendanceSessionResponse": {
        "properties": {
          "sessionDate": {
            "type": "string",
            "title": "Sessiondate"
          },
          "shift": {
            "type": "string",
            "title": "Shift"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "in_progress"
          },
          "supervisorId": {
            "type": "string",
            "title": "Supervisorid"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": " Id"
          },
          "finalizedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizedat"
          },
          "finalizedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizedby"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "sessionDate",
          "shift",
          "supervisorId"
        ],
        "title": "AttendanceSessionResponse"
      },
...
      "EmployeeCreate": {
        "properties": {
          "employeeId": {
            "type": "string",
            "title": "Employeeid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "photoPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photopath"
          },
          "displayOrder": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayorder"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "contractor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contractor"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "testField": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Testfield"
          }
        },
        "type": "object",
        "required": [
          "employeeId",
          "name",
          "category"
        ],
        "title": "EmployeeCreate"
      },
      "EmployeeResponse": {
        "properties": {
          "employeeId": {
            "type": "string",
            "title": "Employeeid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "photoPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photopath"
          },
          "displayOrder": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayorder"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "contractor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contractor"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "testField": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Testfield"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "employeeId",
          "name",
          "category"
        ],
        "title": "EmployeeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehicleAssignmentResponse": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "employeeId": {
            "type": "string",
            "title": "Employeeid"
          },
          "vehicleId": {
            "type": "string",
            "title": "Vehicleid"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Assignedat"
          },
          "assignedBy": {
            "type": "string",
            "title": "Assignedby"
          },
          "unassignedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unassignedat"
          },
          "unassignedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unassignedby"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "employeeId",
          "vehicleId",
          "assignedBy"
        ],
        "title": "VehicleAssignmentResponse"
      },
      "VehicleCreate": {
        "properties": {
          "vehicleNumber": {
            "type": "string",
            "title": "Vehiclenumber"
          },
          "vehicleType": {
            "type": "string",
            "title": "Vehicletype"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "available"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "vehicleNumber",
          "vehicleType"
        ],
        "title": "VehicleCreate"
      },
      "VehicleResponse": {
        "properties": {
          "vehicleNumber": {
            "type": "string",
            "title": "Vehiclenumber"
          },
          "vehicleType": {
            "type": "string",
            "title": "Vehicletype"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "available"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "vehicleNumber",
          "vehicleType"
        ],
        "title": "VehicleResponse"
      },
      "VehicleUpdate": {
        "properties": {
          "vehicleNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiclenumber"
          },
          "vehicleType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicletype"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "type": "object",
        "title": "VehicleUpdate"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}